fixing tests 1
