+++ src/template/pt_template.rs
@@ -24,7 +24,7 @@ use crate::{
 use super::{header_duplet_row::HeaderDupletRow, header_index::HeaderIndexer, operations::Operation};
 
 /// Phetools can be used to curate cases with Mendelian disease or with melded phenotypes
-#[derive(PartialEq)]
+#[derive(Debug,PartialEq)]
 pub enum TemplateType {
     Mendelian,
     Melded,
@@ -543,4 +543,185 @@ impl PheToolsTemplate {
         summary.insert("n_ppkt".to_string(), format!("{}", ppkt_n));
         summary
     }
+}
+
+
+#[cfg(test)]
+mod test {
+    use crate::{error::Error, header::{header_duplet::HeaderDupletItem, hpo_term_duplet::HpoTermDuplet}};
+    use lazy_static::lazy_static;
+    use ontolius::{io::OntologyLoaderBuilder, ontology::csr::MinimalCsrOntology};
+    use polars::io::SerReader;
+    use super::*;
+    use std::{fs::File, io::BufReader};
+    use rstest::{fixture, rstest};
+    use flate2::bufread::GzDecoder;
+
+    #[fixture]
+    fn hpo() -> Arc<FullCsrOntology> {
+        let path = "resources/hp.v2025-03-03.json.gz";
+        let reader = GzDecoder::new(BufReader::new(File::open(path).unwrap()));
+        let loader = OntologyLoaderBuilder::new().obographs_parser().build();
+        let hpo = loader.load_from_read(reader).unwrap();
+        Arc::new(hpo)
+    }
+
+
+
+    #[fixture]
+    fn row1() -> Vec<String> 
+    {
+        let row: Vec<&str> = vec![
+            "PMID", "title", "individual_id", "comment", "disease_id", "disease_label", "HGNC_id",	"gene_symbol", 
+            "transcript", "allele_1", "allele_2", "variant.comment", "age_of_onset", "age_at_last_encounter", 
+            "deceased", "sex", "HPO",	"Clinodactyly of the 5th finger", "Hallux valgus",	"Short 1st metacarpal", 
+            "Ectopic ossification in muscle tissue", "Long hallux", "Pain", "Short thumb"
+        ];
+        row.into_iter().map(|s| s.to_owned()).collect()
+    }
+
+    #[fixture]
+    fn row2() -> Vec<String> 
+    {
+        let row: Vec<&str> = vec![
+            "CURIE", "str", "str", "optional", "CURIE", "str", "CURIE", "str", "str", "str", "str", "optional", "age", "age", "yes/no/na", "M:F:O:U", "na",
+            "HP:0004209", "HP:0001822", "HP:0010034", "HP:0011987", "HP:0001847", "HP:0012531", "HP:0009778"];
+        row.into_iter().map(|s| s.to_owned()).collect()
+    }
+
+    #[fixture]
+    fn row3() -> Vec<String> {
+        let row: Vec<&str> =  vec![
+            "PMID:29482508", "Difficult diagnosis and genetic analysis of fibrodysplasia ossificans progressiva: a case report", "current case", "", 
+            "OMIM:135100", "Fibrodysplasia ossificans progressiva", "HGNC:171", "ACVR1", 
+            "NM_001111067.4", "c.617G>A", "na", "NP_001104537.1:p.(Arg206His)", 
+            "P9Y", "P16Y", "no", "M", "na", "na", "P16Y", "na", "P16Y", "P16Y", "P16Y", "na"];
+        row.into_iter().map(|s| s.to_owned()).collect()
+    }
+
+
+    #[fixture]
+    fn original_matrix(row1: Vec<String>, row2: Vec<String>, row3: Vec<String>)  -> Vec<Vec<String>> {
+        let mut rows = Vec::with_capacity(3);
+        rows.push(row1);
+        rows.push(row2);
+        rows.push(row3);
+        rows
+    }
+
+    /// Make sure that our test matrix is valid before we start changing fields to check if we pick up errors
+    #[rstest]
+    fn test_factory_valid_input(
+        original_matrix: Vec<Vec<String>>, 
+        hpo: Arc<FullCsrOntology>) {
+        let factory = PheToolsTemplate::from_mendelian_template(original_matrix, hpo);
+        assert!(factory.is_ok());
+    }
+
+
+    #[rstest]
+    fn test_malformed_hpo_label(mut original_matrix: Vec<Vec<String>>, hpo: Arc<FullCsrOntology>) {
+        // "Hallux valgus" has extra white space
+        original_matrix[0][19] = "Hallux  valgus".to_string(); 
+        let factory = PheToolsTemplate::from_mendelian_template(original_matrix, hpo);
+        assert!(&factory.is_err());
+        assert!(matches!(&factory, Err(Error::TermError { .. })));
+        let err = factory.err().unwrap();
+        let err_msg = err.to_string();
+        let expected = "HPO Term HP:0010034 with malformed label 'Hallux  valgus' instead of 'Short 1st metacarpal'";
+        assert_eq!(expected, err_msg);
+    }
+
+
+   
+    /// Test that we detect errors in labels of headings
+    #[rstest]
+    #[case(0, "PMI", "PMID")]
+    #[case(1, "title ", "title")]
+    #[case(1, " title ", "title")]
+    #[case(1, "titl", "title")]
+    #[case(2, "individual", "individual_id")]
+    #[case(3, "comm", "comment")]
+    #[case(4, "disease_i", "disease_id")]
+    #[case(5, "diseaselabel", "disease_label")]
+    #[case(6, "hgnc", "HGNC_id")]
+    #[case(7, "symbol", "gene_symbol")]
+    #[case(8, "tx", "transcript")]
+    #[case(9, "allel1", "allele_1")]
+    #[case(910, "allele2", "allele_2")]
+    #[case(11, "vcomment", "variant.comment")]
+    #[case(12, "age", "age_of_onset")]
+    #[case(13, "age_last_counter", "age_at_last_encounter")]
+    #[case(14, "deceasd", "deceased")]
+    #[case(15, "sexcolumn", "sex")]
+    #[case(16, "", "HPO")]
+    fn test_malformed_title_row(
+        mut original_matrix: Vec<Vec<String>>, 
+        hpo: Arc<FullCsrOntology>, 
+        #[case] idx: usize, 
+        #[case] label: &str,
+        #[case] expected_label: &str) {
+        // Test that we catch malformed labels for the first row
+        original_matrix[0][idx] = label.to_string(); 
+        let pt_template = PheToolsTemplate::from_mendelian_template(original_matrix, hpo);
+        assert!(&pt_template.is_err());
+        assert!(matches!(&pt_template, Err(Error::HeaderError { .. })));
+        let err = pt_template.err().unwrap();
+        let err_msg = err.to_string();
+        let expected = format!("Malformed header: Expected '{}' but got '{}'", expected_label, label );
+        assert_eq!(expected, err_msg); 
+    }
+
+    // test malformed entries
+    // we change entries in the third row (which is the first and only data row)
+    // and introduce typical potential errors
+    #[rstest]
+    #[case(0, "PMID29482508", "Invalid CURIE with no colon: 'PMID29482508'")]
+    #[case(0, "PMID: 29482508", "Contains stray whitespace: 'PMID: 29482508'")]
+    #[case(1, "", "Value must not be empty")]
+    #[case(1, "Difficult diagnosis and genetic analysis of fibrodysplasia ossificans progressiva: a case report ", 
+        "Trailing whitespace in 'Difficult diagnosis and genetic analysis of fibrodysplasia ossificans progressiva: a case report '")]
+    #[case(2, "individual(1)", "Forbidden character '(' found in label 'individual(1)'")]
+    #[case(2, " individual A", "Leading whitespace in ' individual A'")]
+    #[case(4, "MIM:135100", "Disease id has invalid prefix: 'MIM:135100'")]
+    #[case(4, "OMIM: 135100", "Contains stray whitespace: 'OMIM: 135100'")]
+    #[case(4, "OMIM:13510", "OMIM identifiers must have 6 digits: 'OMIM:13510'")]
+    #[case(5, "Fibrodysplasia ossificans progressiva ", "Trailing whitespace in 'Fibrodysplasia ossificans progressiva '")]
+    #[case(6, "HGNC:171 ", "Contains stray whitespace: 'HGNC:171 '")]
+    #[case(6, "HGNC171", "Invalid CURIE with no colon: 'HGNC171'")]
+    #[case(6, "HGNG:171", "HGNC id has invalid prefix: 'HGNG:171'")]
+    #[case(7, "ACVR1 ", "Trailing whitespace in 'ACVR1 '")]
+    #[case(8, "NM_001111067", "Transcript 'NM_001111067' is missing a version")]
+    #[case(9, "617G>A", "Malformed allele '617G>A'")]
+    #[case(10, "", "Value must not be empty")]
+    #[case(12, "P2", "Malformed age_of_onset 'P2'")]
+    #[case(13, "Adultonset", "Malformed age_at_last_encounter 'Adultonset'")]
+    #[case(14, "?", "Malformed deceased entry: '?'")]
+    #[case(14, "alive", "Malformed deceased entry: 'alive'")]
+    #[case(15, "male", "Malformed entry in sex field: 'male'")]
+    #[case(15, "f", "Malformed entry in sex field: 'f'")]
+    #[case(18, "Observed", "Malformed entry for Ectopic ossification in muscle tissue (HP:0011987): 'Observed'")]
+    #[case(18, "yes", "Malformed entry for Ectopic ossification in muscle tissue (HP:0011987): 'yes'")]
+    #[case(18, "exc.", "Malformed entry for Ectopic ossification in muscle tissue (HP:0011987): 'exc.'")]
+      fn test_malformed_entry(
+        mut original_matrix: Vec<Vec<String>>, 
+        hpo: Arc<FullCsrOntology>, 
+        #[case] idx: usize, 
+        #[case] entry: &str,
+        #[case] error_msg: &str) 
+    {
+        original_matrix[2][idx] = entry.to_string();
+        let factory = PheToolsTemplate::from_mendelian_template(original_matrix, hpo);
+        assert!(factory.is_ok());
+        let factory = factory.unwrap();
+       /*  let templates = factory.get_templates().unwrap();
+        assert_eq!(1, templates.len());
+        let itemplate = &templates[0];
+        let result = itemplate.qc_check();
+        assert!(result.is_err());
+        let err = result.unwrap_err();
+        assert_eq!(error_msg, err.to_string());*/
+    }
+
+
 }
